updates, new fucntions ...
